Combine Http401UnauthorizedEntryPoint with StormpathAuthenticationEntryPoint
